requirement.txt
flask-restful
opencv-python
mysql-connector-python
